+++ service/src/main/java/com/nosota/mwallet/dto/TransactionDTO.java
@@ -1,5 +1,6 @@
 package com.nosota.mwallet.dto;
 
+import com.nosota.mwallet.model.Transaction;
 import com.nosota.mwallet.model.TransactionStatus;
 import com.nosota.mwallet.model.TransactionType;