+++ service/src/main/java/com/nosota/mwallet/repository/TransactionSnapshotRepository.java
@@ -6,6 +6,10 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
+import java.time.LocalDate;
+import java.time.LocalDateTime;
+import java.util.List;
+
 @Repository
 public interface TransactionSnapshotRepository extends JpaRepository<TransactionSnapshot, Integer> {
     /**
@@ -17,12 +21,30 @@ public interface TransactionSnapshotRepository extends JpaRepository<Transaction
     @Query("SELECT COALESCE(SUM(t.amount), 0) FROM TransactionSnapshot t WHERE t.walletId = :walletId AND t.status = 'HOLD'")
     Long getHoldBalanceForWallet(@Param("walletId") Integer walletId);
 
-    /**
-     * Retrieves the sum of all RESERVED transaction amounts for the specified wallet.
-     *
-     * @param walletId The unique identifier (ID) of the wallet.
-     * @return The total of RESERVED amounts. If no transactions are found, returns 0.
-     */
     @Query("SELECT COALESCE(SUM(t.amount), 0) FROM TransactionSnapshot t WHERE t.walletId = :walletId AND t.status = 'RESERVED'")
     Long getReservedBalanceForWallet(@Param("walletId") Integer walletId);
+
+    @Query("SELECT ts FROM TransactionSnapshot ts WHERE ts.walletId = :walletId AND ts.type = 'CREDIT' AND ts.status = 'CONFIRMED' " +
+            "AND ts.confirmRejectTimestamp BETWEEN :startOfDay AND :date")
+    List<TransactionSnapshot> findDailyCreditSnapshotOperations(@Param("walletId") Integer walletId,
+                                                                @Param("startOfDay") LocalDateTime startOfDay,
+                                                                @Param("date") LocalDateTime date);
+
+    @Query("SELECT ts FROM TransactionSnapshot ts WHERE ts.walletId = :walletId AND ts.type = 'DEBIT' AND ts.status = 'CONFIRMED' " +
+            "AND ts.confirmRejectTimestamp BETWEEN :startOfDay AND :date")
+    List<TransactionSnapshot> findDailyDebitSnapshotOperations(@Param("walletId") Integer walletId,
+                                                               @Param("startOfDay") LocalDateTime startOfDay,
+                                                               @Param("date") LocalDateTime date);
+
+    @Query("SELECT ts FROM TransactionSnapshot ts WHERE ts.walletId = :walletId AND ts.type = 'CREDIT' AND ts.status = 'CONFIRMED' " +
+            "AND ts.confirmRejectTimestamp BETWEEN :fromDate AND :toDate")
+    List<TransactionSnapshot> findCreditSnapshotOperationsInRange(@Param("walletId") Integer walletId,
+                                                                  @Param("fromDate") LocalDateTime fromDate,
+                                                                  @Param("toDate") LocalDateTime toDate);
+
+    @Query("SELECT ts FROM TransactionSnapshot ts WHERE ts.walletId = :walletId AND ts.type = 'DEBIT' AND ts.status = 'CONFIRMED' " +
+            "AND ts.confirmRejectTimestamp BETWEEN :fromDate AND :toDate")
+    List<TransactionSnapshot> findDebitSnapshotOperationsInRange(@Param("walletId") Integer walletId,
+                                                                 @Param("fromDate") LocalDateTime fromDate,
+                                                                 @Param("toDate") LocalDateTime toDate);
 }