+++ service/src/test/java/com/nosota/mwallet/TransactionArchiveTest.java
@@ -26,9 +26,6 @@ public class TransactionArchiveTest {
     private TransactionSnapshotService transactionSnapshotService;
 
     @Autowired
-    private WalletService walletService;
-
-    @Autowired
     private WalletBalanceService walletBalanceService;
 
     @Autowired