+++ service/src/test/java/com/nosota/mwallet/TransactionSnapshotTest.java
@@ -25,9 +25,6 @@ public class TransactionSnapshotTest {
     private TransactionSnapshotService transactionSnapshotService;
 
     @Autowired
-    private WalletService walletService;
-
-    @Autowired
     private WalletBalanceService walletBalanceService;
 
     @Autowired